Реализовано разделение информации пользователей(счета, категории, операции)

+++ backend/controllers/account.js
@@ -1,3 +1,4 @@
+const mongoose = require('mongoose');
 const Account = require('../models/Account');
 const Transaction = require('../models/Transaction');
 
@@ -9,8 +10,18 @@ async function addAccount(account) {
 }
 
 // edit
-async function editAccount(id, account) {
-  const newAccount = await Account.findByIdAndUpdate(id, account, {
+async function editAccount(user, id, accountData) {
+  const account = await Account.findById(id);
+
+  if (!account) {
+    throw new Error('Account not found');
+  }
+
+  if (user !== account.user.toString()) {
+    throw new Error('Account access error');
+  }
+
+  const newAccount = await Account.findByIdAndUpdate(id, accountData, {
     returnDocument: 'after',
   });
 
@@ -18,22 +29,44 @@ async function editAccount(id, account) {
 }
 
 // delete
-async function deleteAccount(id) {
+async function deleteAccount(user, id) {
+  const account = await Account.findById(id);
+
+  if (!account) {
+    throw new Error('Account not found');
+  }
+
+  if (user !== account.user.toString()) {
+    throw new Error('Account access error');
+  }
+
   await Transaction.deleteMany({ account: id });
 
   return Account.deleteOne({ _id: id });
 }
 
 // get list with search and pagination
-async function getAccounts() {
-  const accounts = await Account.find();
+async function getAccounts(user) {
+  const userObjectId = new mongoose.Types.ObjectId(user);
+
+  const accounts = await Account.find({ user: { $in: userObjectId } });
 
   return accounts.sort((a, b) => b.balance - a.balance);
 }
 
 // get item
-function getAccount(id) {
-  return Account.findById(id);
+async function getAccount(user, id) {
+  const account = await Account.findById(id);
+
+  if (!account) {
+    throw new Error('Account not found');
+  }
+
+  if (user !== account.user.toString()) {
+    throw new Error('Account access error');
+  }
+
+  return account;
 }
 
 module.exports = {
